+++ matholymp/roundupreg/templating.py
@@ -47,7 +47,8 @@ __all__ = ['people_from_country_internal', 'people_from_country',
            'register_templating_utils', 'show_prereg_sidebar',
            'show_prereg_reminder', 'bulk_csv_contents',
            'show_bulk_csv_country', 'show_bulk_csv_country_link_from_code',
-           'show_bulk_csv_person', 'bulk_zip_ref', 'required_user_fields']
+           'show_bulk_csv_person', 'bulk_zip_ref', 'required_user_fields',
+           'registration_enabled']
 
 import base64
 import datetime
@@ -77,7 +78,7 @@ from matholymp.roundupreg.roundupsitegen import RoundupSiteGenerator
 from matholymp.roundupreg.rounduputil import person_date_of_birth, \
     contestant_age, person_is_contestant, contestant_code, pn_score, \
     scores_final, any_scores_missing, country_has_contestants, \
-    valid_country_problem
+    valid_country_problem, registration_enabled
 from matholymp.roundupreg.staticsite import static_site_event_group
 
 
@@ -752,3 +753,4 @@ def register_templating_utils(instance):
     instance.registerUtil('show_bulk_csv_person', show_bulk_csv_person)
     instance.registerUtil('bulk_zip_ref', bulk_zip_ref)
     instance.registerUtil('required_user_fields', required_user_fields)
+    instance.registerUtil('registration_enabled', registration_enabled)